requirements.txt
Pillow
qrcode
pycryptodome